change: move ParsedYaml instances into schema files
