+++ src/schema/QuestOpenOverrides.ts
@@ -1,5 +1,10 @@
+import path from "path";
 import { z } from "zod";
+import { Log } from "~/utils/log";
+import { ParsedYaml } from "./ParsedYaml";
 import { zDate } from "./zDate";
+
+export type QuestOpenOverrides = PartialDataMap<number>;
 
 const _zOverrideKey = z
   .string()
@@ -8,23 +13,31 @@ const _zOverrideKey = z
     "openTimeOverrides constants may only use uppercase letters and underscores and must start with a letter."
   );
 const _zTimeStamp = z.number().min(1498449600);
+export const questOpenFilePath = path.join(
+  process.cwd(),
+  "assets/data/upgrades/openTimeOverrides.yml"
+);
 
-export const QuestOpenOverridesSchema = z
-  .object({
-    constants: z.record(
-      _zOverrideKey,
-      z.union([_zTimeStamp, zDate, z.undefined()])
-    ),
-    overrides: z.record(
-      z
-        .string()
-        .regex(/^\d+$/)
-        .transform(val => Number(val)),
-      z.union([_zOverrideKey, _zTimeStamp, zDate])
-    )
-  })
-  .transform(({ overrides, constants }, ctx) => {
-    const overridesMapped: PartialDataMap<number> = {};
+export const QuestOpenOverridesSchema = z.object({
+  constants: z.record(
+    _zOverrideKey,
+    z.union([_zTimeStamp, zDate, z.undefined()])
+  ),
+  overrides: z.record(
+    z
+      .string()
+      .regex(/^\d+$/)
+      .transform(val => Number(val)),
+    z.union([_zOverrideKey, _zTimeStamp, zDate])
+  )
+});
+
+export const QuestOpenOverridesFile = new ParsedYaml({
+  name: "Quest Open Overrides",
+  limitPath: questOpenFilePath,
+  schema: QuestOpenOverridesSchema,
+  transform: ({ overrides, constants }) => {
+    const overridesMapped: QuestOpenOverrides = {};
     const entries = Object.entries(overrides) as [
       `${number}`,
       string | number
@@ -38,22 +51,18 @@ export const QuestOpenOverridesSchema = z
 
       const mappedVal = constants[val];
       if (typeof mappedVal == "undefined") {
-        ctx.addIssue({
-          code: z.ZodIssueCode.custom,
-          message: "Unrecognized constant or date",
-          path: ["overrides", key]
-        });
-
-        return z.NEVER;
+        Log.throw(`Unrecognized constant or date for quest id ${key}`);
       }
 
       overridesMapped[key] = mappedVal;
     }
 
     return overridesMapped;
-  });
+  }
+});
 
-export type QuestOpenOverrides = {
-  in: z.input<typeof QuestOpenOverridesSchema>;
-  out: z.output<typeof QuestOpenOverridesSchema>;
-};
+export async function getQuestOpenOverrides() {
+  const res = await QuestOpenOverridesFile.readFile(questOpenFilePath);
+  if (!res.success) throw res.error;
+  return res.data;
+}
